refactor: consolidate Bailing MTP code
